+++ src/Symfony/Component/HttpClient/Tests/Response/JsonMockResponseTest.php
@@ -85,4 +85,25 @@ final class JsonMockResponseTest extends TestCase
             ['application/problem+json', ['x-foo' => 'ccc', 'content-type' => 'application/problem+json']],
         ];
     }
+
+    public function testFromFile()
+    {
+        $client = new MockHttpClient(JsonMockResponse::fromFile(__DIR__.'/Fixtures/response.json'));
+        $response = $client->request('GET', 'https://symfony.com');
+
+        $this->assertSame([
+            'foo' => 'bar',
+        ], $response->toArray());
+        $this->assertSame('application/json', $response->getHeaders()['content-type'][0]);
+    }
+
+    public function testFromFileWithInvalidJson()
+    {
+        $path = __DIR__.'/Fixtures/invalid_json.json';
+
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage(sprintf('File "%s" does not contain valid JSON.', $path));
+
+        JsonMockResponse::fromFile($path);
+    }
 }